Make toolbars separated for each "FileNavigator"
