+++ client-react/src/client/components/FileManager/FileManager.less
@@ -9,7 +9,6 @@
 .oc-fm--file-manager__navigators {
   flex: 1;
   display: flex;
-  border-top: 1px solid #e5e5e5;
 }
 
 .oc-fm--file-manager__navigator {
@@ -19,9 +18,4 @@
   &:last-child {
     border-right: none;
   }
-}
-
-.oc-fm--file-manager__toolbar {
-  padding: 6px 8px;
-  box-shadow: 0 -2px 8px rgba(0,0,0,0.12);
 }
