+++ client-react/src/client/components/FileNavigator/FileNavigator.less
@@ -7,14 +7,20 @@
   flex-direction: column;
 }
 
+.oc-fm--file-navigator__toolbar {
+  padding: 6px 8px;
+  box-shadow: 0 -2px 8px rgba(0,0,0,0.12);
+}
+
 .oc-fm--file-navigator__view {
   flex: 1;
-  position: relative;
   overflow: hidden;
+  position: static !important; // XXX - Fix for ListView Autosizer append inline styles to parent element
 }
 
 .oc-fm--file-navigator__location-bar {
   background: #f5f5f5;
+  padding: 0 8px;
 }
 
 .oc-fm--file-navigator__notifications {
